At instead of [] operator

+++ src/SeisSol/ReceiverDB.h
@@ -18,10 +18,10 @@ class ReceiverDB {
   size_t numberOfReceivers(size_t fsn = 1) const;
 
   private:
-  std::map<size_t, std::map<int, SeisSol::Receiver>> receivers;
+  std::map<int, std::map<int, SeisSol::Receiver>> receivers;
   const IO::ReceiverReader receiverReader;
   std::string directory;
   std::string prefix;
-};
+};  
 
 } // namespace SeisSol
